+++ services.php
@@ -1,11 +1,15 @@
 <?php
 
+use Dhii\Output\PlaceholderTemplate;
+use Dhii\Output\PlaceholderTemplateFactory;
+use Dhii\Output\TemplateFactoryInterface;
 use Psr\Container\ContainerInterface;
 use RebelCode\EddBookings\Cart\BookingPriceEvaluator;
 use RebelCode\EddBookings\Cart\BookingValueAwareFactory;
 use RebelCode\EddBookings\Cart\Module\AddBookingToCartHandler;
 use RebelCode\EddBookings\Cart\Module\FilterCartItemPriceHandler;
 use RebelCode\EddBookings\Cart\Module\RemoveBookingFromCartHandler;
+use RebelCode\EddBookings\Cart\Module\RenderCartBookingInfoHandler;
 use RebelCode\EddBookings\Cart\Module\SubmitBookingOnPaymentHandler;
 use RebelCode\EddBookings\Cart\Module\ValidateCartBookingHandler;
 
@@ -27,7 +31,7 @@ return [
     'eddbk_add_booking_to_cart_handler' => function (ContainerInterface $c) {
         return new AddBookingToCartHandler(
             $c->get('edd_cart'),
-            $c->get('edd_cart_config/items')
+            $c->get('eddbk_cart/cart_items')
         );
     },
 
@@ -42,7 +46,7 @@ return [
             $c->get('bookings_select_rm'),
             $c->get('bookings_delete_rm'),
             $c->get('sql_expression_builder'),
-            $c->get('edd_cart_config/items')
+            $c->get('eddbk_cart/cart_items')
         );
     },
 
@@ -57,7 +61,7 @@ return [
             $c->get('bookings_select_rm'),
             $c->get('bookings_update_rm'),
             $c->get('sql_expression_builder'),
-            $c->get('edd_cart_config/items')
+            $c->get('eddbk_cart/cart_items')
         );
     },
 
@@ -72,11 +76,11 @@ return [
             $c->get('booking_transitioner'),
             $c->get('bookings_select_rm'),
             $c->get('sql_expression_builder'),
-            $c->get('edd_cart_config/items')
+            $c->get('eddbk_cart/cart_items')
         );
     },
 
-    /**
+    /*
      * The handler that filters cart item prices.
      *
      * @since [*next-version*]
@@ -87,11 +91,54 @@ return [
             $c->get('eddbk_booking_price_evaluator'),
             $c->get('eddbk_booking_value_aware_factory'),
             $c->get('sql_expression_builder'),
-            $c->get('edd_cart_config/items')
+            $c->get('eddbk_cart/cart_items')
         );
     },
 
-    /**
+    /*
+     * The handler that renders booking information in the EDD cart.
+     *
+     * @since [*next-version*]
+     */
+    'eddbk_render_cart_booking_info_handler' => function (ContainerInterface $c) {
+        return new RenderCartBookingInfoHandler(
+            $c->get('eddbk_cart_booking_info_template'),
+            $c->get('bookings_select_rm'),
+            $c->get('sql_expression_builder'),
+            $c->get('eddbk_cart/cart_items')
+        );
+    },
+
+    /*
+     * The factory used to create templates used in this module.
+     *
+     * @since [*next-version*]
+     */
+    'eddbk_cart_template_factory' => function (ContainerInterface $c) {
+        return new PlaceholderTemplateFactory(
+            'Dhii\Output\PlaceholderTemplate',
+            $c->get('eddbk_cart/templates/token_start'),
+            $c->get('eddbk_cart/templates/token_end'),
+            $c->get('eddbk_cart/templates/token_default')
+        );
+    },
+
+    /*
+     * The template for booking info in the EDD cart.
+     *
+     * @since [*next-version*]
+     */
+    'eddbk_cart_booking_info_template' => function (ContainerInterface $c) {
+        $templateFile = $c->get('eddbk_cart/cart_items/templates/booking_info');
+        $templatePath = EDDBK_CART_MODULE_TEMPLATES_DIR . DIRECTORY_SEPARATOR . $templateFile;
+        $template = file_get_contents($templatePath);
+
+        return $c->get('eddbk_cart_template_factory')->make([
+            TemplateFactoryInterface::K_TEMPLATE => $template
+        ]);
+    },
+
+    /*
      * The booking price evaluator.
      *
      * @since [*next-version*]
@@ -103,7 +150,7 @@ return [
         );
     },
 
-    /**
+    /*
      * The factory that creates booking value-aware instances.
      *
      * @since [*next-version*]