+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [[*next-version*]] - YYYY-MM-DD
+### Added
+- Booking info is now shown in the EDD cart for cart items that correspond to bookings.
 
 ## [0.1-alpha1] - 2018-06-04
 Initial version.